requirements.txt
python-ldap==2.4.18

# for tests
python-ldap-test==0.2.0
nose==1.3.4
rednose==0.4.1
mock==1.0.1
git+https://github.com/sttts/mockldap.git@egg=mockldap
funcparserlib==0.3.6
